refact: use GetAsString
